requirements.txt
flask
pydantic
python-dotenv
openai
langchain
requests